Fix: phpstan

+++ sfapp/src/Service/AlertManager.php
@@ -155,7 +155,9 @@ class AlertManager
     {
         $alerts = $this->alertRepository->findAlertsByRoom($room);
         foreach ($alerts as $alert) {
-            $this->entityManager->remove($alert);
+            if ($alert instanceof Alert) {
+                $this->entityManager->remove($alert);
+            }
         }
         $this->entityManager->flush();
     }
